% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{batch_correction}
\alias{batch_correction}
\title{Batch_correction}
\usage{
batch_correction(data, method = "ComBat")
}
\arguments{
\item{data}{Normalized single-cell metabolomics data matrix}

\item{method}{ComBat,Limma}
}
\value{
Corrected single-cell metabolomics data matrix
}
\description{
Batch_correction
}
\examples{

normalized_data <- cbind(matrix(rnorm(40), nrow = 10), matrix(rnorm(60), nrow = 10))

batch_info <- factor(sample(1:2, 10, replace = TRUE))
normalized_data <- cbind(normalized_data, batch_info)

corrected_data <- batch_correction(t(normalized_data), method = "ComBat")

}
