% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Functional_classification}
\usage{
classification(data2, label_col, folds = 5, method)
}
\arguments{
\item{data2}{Matrix of preprocessed single-cell metabolomics data}

\item{label_col}{A vector indicating the group labels (e.g., 0 and 1) for each sample in the data, used to distinguish different groups for the statistical test.}

\item{folds}{The number of folds to use for cross - validation. Default is 5, which means the data will be split into 5 subsets for cross - validation purposes.}

\item{method}{The classification method to be used. Valid options include "AdaBoost", "Bagging", "Decision Trees", etc.}
}
\value{
A list containing two elements. The first element is a data frame with the true and predicted labels for each fold of the cross - validation. The second element is a data frame with the top 10 most important features (as determined by the selected classification method) for each fold.
}
\description{
Functional_classification
}
\examples{
\donttest{
}






}
