% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classplot.R
\name{classification_plots}
\alias{classification_plots}
\title{Classification_plots}
\usage{
classification_plots(final_res_output)
}
\arguments{
\item{final_res_output}{}
}
\value{
Receiver operating characteristic (ROC) curve and area under the curve (AUC) value
}
\description{
Classification_plots
}
\examples{
\donttest{
}
}
