% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Filtering}
\usage{
filtering(data, percentage)
}
\arguments{
\item{data}{Single-cell metabolomics data matrix}

\item{percentage}{Percent of Missing Values}
}
\value{
Filtered single-cell metabolomics data matrix
}
\description{
Filtering
}
\examples{

set.seed(123)
data <- matrix(rnorm(100), ncol = 10)

data <- cbind(matrix(sample(1:10, 40, replace = TRUE), ncol = 4), data)

filted_data <- filtering(data, 0.2)

}
