% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{imputation}
\alias{imputation}
\title{Imputation}
\usage{
imputation(
  data,
  method = "KNN",
  k = 10,
  rowmax = 0.5,
  colmax = 0.8,
  maxp = 1500
)
}
\arguments{
\item{data}{Filtered single-cell metabolomics data matrix}

\item{method}{knn,1/5 of minimum positive value}

\item{k}{k represents the number of nearest neighbors considered for imputing missing values in the KNN algorithm. A small k may make the imputation unstable due to outliers, while a large k may introduce bias by including dissimilar samples.}

\item{rowmax}{rowmax is a proportion indicating the maximum allowable proportion of missing values per row. Rows with a higher proportion won't be used for calculating neighbors, avoiding the influence of rows with excessive missing values.}

\item{colmax}{colmax is a proportion defining the maximum allowable proportion of missing values per column. Columns exceeding this proportion won't be imputed, preventing inaccurate imputation for columns with too many missing values.}

\item{maxp}{maxp is an integer that limits the maximum number of features considered when calculating neighbors. A small maxp may overlook important features, while a large one increases computational complexity and may introduce noise.}
}
\value{
Imputed single-cell metabolomics data matrix
}
\description{
Imputation
}
\examples{

set.seed(123)
filted_data <- matrix(rnorm(100), ncol = 10)

filted_data <- cbind(matrix(sample(1:10, 40, replace = TRUE), ncol = 4), filted_data)

imputed_data <- imputation(filted_data, method = "1/5 of minimum positive value")

imputed_data <- imputation(filted_data, method = "KNN")

}
