% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolicclassification.R
\name{metabolic_classification}
\alias{metabolic_classification}
\title{Functional_classification}
\usage{
metabolic_classification(data2, labels, folds = 5, method)
}
\arguments{
\item{labels}{A vector of class labels corresponding to each row in the `data` matrix. These labels should represent the different classes or groups for the classification task.}

\item{folds}{The number of folds to use for cross - validation. Default is 5, which means the data will be split into 5 subsets for cross - validation purposes.}

\item{method}{The classification method to be used. Valid options include "AdaBoost", "Bagging", "Decision Trees", etc.}

\item{data}{Matrix of preprocessed single-cell metabolomics data}
}
\value{
A list containing two elements. The first element is a data frame with the true and predicted labels for each fold of the cross - validation. The second element is a data frame with the top 10 most important features (as determined by the selected classification method) for each fold.
}
\description{
Functional_classification
}
\examples{
\donttest{
}






}
