% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolicfeature.R
\name{metabolic_differential}
\alias{metabolic_differential}
\title{Metabolic_differential}
\usage{
metabolic_differential(
  data,
  labels,
  method = c("t_test", "ANOVA", "FC", "PLS-DA", "OPLS-DA", "RF_RFE", "Kruskal_Wallis",
    "svmrfeFeatureRanking"),
  typeFDR = "BH",
  algo = "t",
  q = 0.05,
  ntree = 500,
  ntreeIterat = 300,
  vars.drop.frac = 0.2,
  cValue = 50,
  Scaled = FALSE
)
}
\arguments{
\item{data}{Matrix of preprocessed single-cell metabolomics data}

\item{labels}{A vector indicating the group labels (e.g., 0 and 1) for each sample in the data, used to distinguish different groups for the statistical test.}

\item{method}{"t_test", "ANOVA", "FC", "PLS-DA", "OPLS-DA", "RF_RFE", "Kruskal_Wallis", "svmrfeFeatureRanking"}

\item{typeFDR}{The method for false discovery rate (FDR) adjustment of p - values, with "BH" (Benjamini - Hochberg) as a common option.}

\item{algo}{The statistical test algorithm to be used, such as "t" for the t - test, to calculate the test statistic.}

\item{q}{The significance level threshold for the adjusted p - values (FDR), typically set to 0.05 to determine significant results.}

\item{ntree}{The number of trees to grow in the random forest model (default 500), affecting model stability and accuracy.}

\item{ntreeIterat}{The number of trees grown in each iteration of the variable - selection algorithm (default 300).}

\item{vars.drop.frac}{vars.drop.frac is the fraction (default 0.2) of variables dropped in each iteration of random forest variable selection, aiding in finding relevant variables.}

\item{cValue}{It is a parameter for the SVM (Support Vector Machine) in the SVM - RFE (Recursive Feature Elimination) process. Specifically, cValue (default is 50) represents the cost parameter in the SVM. A larger cValue makes the SVM try harder to classify all training samples correctly, potentially leading to overfitting, while a smaller value allows for more misclassifications in training to achieve better generalization.}

\item{Scaled}{It is a logical parameter indicating whether the input data should be scaled. When Scaled = TRUE, the input data will be scaled; when Scaled = FALSE (default), no scaling operation will be performed on the input data. Scaling can sometimes improve the performance of SVM by making features on a similar scale.}
}
\value{
A data frame or relevant result object depending on the selected differential analysis method. For example, if "t_test" is selected, it returns a data frame with columns "probeID", "Stat", "RawpValue", "AdjpValue".
}
\description{
Metabolic_differential
}
\examples{
\donttest{
}


}
