% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Normalization}
\usage{
normalization(data, method = "Auto Scaling", sis_column = 2)
}
\arguments{
\item{data}{Transformed single-cell metabolomics data matrix}

\item{method}{'Auto Scaling', 'Cyclic Loess', 'Mean', 'Median', 'MSTUS' , 'SIS'}

\item{sis_column}{sis_column is a parameter in the normalization function that is specifically used when the normalization method is set to "SIS". It indicates the column number in the data matrix that will be used as the reference vector for the internal standard normalization carried out by the SIS function.}
}
\value{
normalized single-cell metabolomics data matrix
}
\description{
Normalization
}
\examples{

transformed_data <- cbind(matrix(rnorm(40), nrow = 10), matrix(rnorm(60), nrow = 10))
normalized_data <- normalization(transformed_data, method = "MSTUS")

}
